+++ ChangeLog.md
@@ -1,3 +1,5 @@
 # Changelog for franz
 
 ## Unreleased changes
+
+* Added `Database.Franz.URI`